+++ src/NeuralNet/Layers/Softmax.php
@@ -137,11 +137,12 @@ class Softmax implements Output, Parametric
         $weights = array_fill(0, $this->width(),
             array_fill(0, $previous->width(), 0.0));
 
-        $r = sqrt(6 / $previous->width());
+        $r = (int) (sqrt(6 / $previous->width()));
 
         for ($i = 0; $i < $this->width(); $i++) {
             for ($j = 0; $j < $previous->width(); $j++) {
-                $weights[$i][$j] = random_int(-$r * 1e8, $r * 1e8) / 1e8;
+                $weights[$i][$j] = random_int((int) (-$r * 1e8),
+                    (int) ($r * 1e8)) / 1e8;
             }
         }
